feat(suite-native): Extend BasicProvider by another providers
